🚧 タイムスタンプを表示できるようにした
