+++ philo/include/philo.h
@@ -6,7 +6,7 @@
 /*   By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/14 19:53:30 by miyuu             #+#    #+#             */
-/*   Updated: 2025/04/14 20:01:44 by miyuu            ###   ########.fr       */
+/*   Updated: 2025/04/15 19:34:10 by miyuu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,6 +23,7 @@
 
 /* macro */
 # define UNIT_CONV 1000
+# define LOG_STATUS "\x1b[32m%ld\x1b[39m %d %s, diff: \x1b[32m%ld\x1b[39m\n"
 
 /* struct */
 typedef struct s_univ_rules
@@ -41,5 +42,9 @@ typedef struct s_thread_arg
 	pthread_mutex_t	*mutex;
 	t_univ_rules	u_rules;
 }	t_thread_arg;
+
+/* function */
+t_univ_rules	init_univ_rules(int argc, char *argv[]);
+long			printf_philo_status(char *status, int n_philo, long last_time);
 
 #endif
